spatie

+++ app/Http/Controllers/Dashboard/CategoryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Category\PutRequest;
 use App\Http\Requests\Category\StoreRequest;
 use App\Models\Category;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class CategoryController extends Controller
@@ -15,6 +16,9 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        if(!Auth::user()->hasPermissionTo('editor.category.index')) {
+            return abort(403);
+        }
         $categories = Category::paginate(2);
         return view('dashboard.category.index', compact('categories'));
         //dd($post->category->title);
@@ -37,6 +41,9 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        if(!Auth::user()->hasPermissionTo('editor.post.create')) {
+            return abort(403);
+        }
         $category = new Category();
         return view('dashboard.category.create', compact('category'));
     }
@@ -46,6 +53,9 @@ class CategoryController extends Controller
      */
     public function store(StoreRequest $request)
     {
+        if(!Auth::user()->hasPermissionTo('editor.post.store')) {
+            return abort(403);
+        }
         $validated = Validator::make($request->all(), [
             'title' => 'required|min:5|min:500',
             'slug' => 'required|min:5|min:500'
@@ -74,6 +84,9 @@ class CategoryController extends Controller
      */
     public function show(Category $category)
     {
+        if(!Auth::user()->hasPermissionTo('editor.post.show')) {
+            return abort(403);
+        }
         return view('dashboard.category.show', compact('category'));
     }
 
@@ -82,6 +95,9 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        if(!Auth::user()->hasPermissionTo('editor.post.update')) {
+            return abort(403);
+        }
         return view('dashboard.category.edit', compact('category'));
     }
 
@@ -90,6 +106,9 @@ class CategoryController extends Controller
      */
     public function update(PutRequest $request, Category $category)
     {
+        if(!Auth::user()->hasPermissionTo('editor.post.update')) {
+            return abort(403);
+        }
         $data = $request->validated();
         $category->update($data);
         return to_route('category.index')->with('status', 'Categoria actualizada');
@@ -100,6 +119,9 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if(!Auth::user()->hasPermissionTo('editor.post.delete')) {
+            return abort(403);
+        }
         $category->delete();
         return to_route('category.index')->with('status', 'Category delete');
     }
